FRONT-01: Setup router
This commit contains the following:
* Installed react-router-dom
* Created router in pages/App.jsx
* Created example routes /, /login
* Created a link to navigate between pages
