+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/daniel/Desktop/frontend-mi-aguila/app/src/index.js":"1","/home/daniel/Desktop/frontend-mi-aguila/app/src/reportWebVitals.js":"2","/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/App.jsx":"3","/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/reducers/index.js":"4","/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/constants.js":"5","/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/actions/global.js":"6","/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/reducers/Global.js":"7"},{"size":629,"mtime":1606579942341,"results":"8","hashOfConfig":"9"},{"size":362,"mtime":1606574924120,"results":"10","hashOfConfig":"9"},{"size":1419,"mtime":1606582451671,"results":"11","hashOfConfig":"9"},{"size":532,"mtime":1606580829925,"results":"12","hashOfConfig":"9"},{"size":324,"mtime":1606577414705,"results":"13","hashOfConfig":"9"},{"size":680,"mtime":1606578842613,"results":"14","hashOfConfig":"9"},{"size":711,"mtime":1606582465661,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1rwbdj7",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/daniel/Desktop/frontend-mi-aguila/app/src/index.js",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/reportWebVitals.js",[],[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/App.jsx",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/reducers/index.js",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/constants.js",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/actions/global.js",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/reducers/Global.js",[]]
+[{"/home/daniel/Desktop/frontend-mi-aguila/app/src/index.js":"1","/home/daniel/Desktop/frontend-mi-aguila/app/src/reportWebVitals.js":"2","/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/App.jsx":"3","/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/reducers/index.js":"4","/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/constants.js":"5","/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/actions/global.js":"6","/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/reducers/Global.js":"7","/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/home/index.js":"8","/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/home/view.jsx":"9","/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/login/view.jsx":"10","/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/login/index.js":"11"},{"size":629,"mtime":1606579942341,"results":"12","hashOfConfig":"13"},{"size":362,"mtime":1606574924120,"results":"14","hashOfConfig":"13"},{"size":802,"mtime":1606584551550,"results":"15","hashOfConfig":"13"},{"size":532,"mtime":1606580829925,"results":"16","hashOfConfig":"13"},{"size":324,"mtime":1606577414705,"results":"17","hashOfConfig":"13"},{"size":680,"mtime":1606578842613,"results":"18","hashOfConfig":"13"},{"size":711,"mtime":1606582465661,"results":"19","hashOfConfig":"13"},{"size":1004,"mtime":1606584165750,"results":"20","hashOfConfig":"13"},{"size":959,"mtime":1606584076856,"results":"21","hashOfConfig":"13"},{"size":1059,"mtime":1606584607374,"results":"22","hashOfConfig":"13"},{"size":1006,"mtime":1606584160160,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1rwbdj7",{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/daniel/Desktop/frontend-mi-aguila/app/src/index.js",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/reportWebVitals.js",[],[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/App.jsx",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/reducers/index.js",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/constants.js",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/actions/global.js",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/redux/reducers/Global.js",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/home/index.js",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/home/view.jsx",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/login/view.jsx",[],"/home/daniel/Desktop/frontend-mi-aguila/app/src/pages/login/index.js",[]]
